+++ scylla/src/execution/iterator.rs
@@ -18,9 +18,9 @@ use tokio::sync::mpsc;
 use tracing::instrument::WithSubscriber;
 
 use crate::cql_to_rust::{FromRow, FromRowError};
+use crate::execution::driver_tracing::RequestSpan;
 use crate::execution::errors::QueryError;
 use crate::execution::ExecutionProfileInner;
-use crate::session::RequestSpan;
 
 use crate::cluster::ClusterData;
 use crate::cluster::{Node, NodeRef};
